# hoist schema require, less work when empty options or hmr === false