--- a/index.js
+++ b/index.js
@@ -14,30 +14,30 @@ module.exports.pitch = function (request) {
 	if (this.cacheable) this.cacheable();
 
   var options = loaderUtils.getOptions(this);
-  if (options) {
-    validateOptions(schema, options, 'Style Loader');
-  } else {
-    options = {};
-  }
 
-	options.hmr = typeof options.hmr === 'undefined' ? true : options.hmr;
-
-	// The variable is needed, because the function should be inlined.
+  // The variable is needed, because the function should be inlined.
 	// If is just stored it in options, JSON.stringify will quote
 	// the function and it would be just a string at runtime
 	var insertInto;
 
-	if (typeof options.insertInto === "function") {
-		insertInto = options.insertInto.toString();
-	}
+  if (options) {
+    validateOptions(schema, options, 'Style Loader');
+    options.hmr = typeof options.hmr === 'undefined' ? true : options.hmr;
 
-	// We need to check if it a string, or variable will be "undefined"
-	// and the loader crashes
-	if (typeof options.insertInto === "string") {
-		insertInto = '"' + options.insertInto + '"';
-	}
+    if (typeof options.insertInto === "function") {
+      insertInto = options.insertInto.toString();
+    }
 
-	var hmr = [
+    // We need to check if it a string, or variable will be "undefined"
+    // and the loader crashes
+    if (typeof options.insertInto === "string") {
+      insertInto = '"' + options.insertInto + '"';
+    }
+  } else {
+    options = { hmr: true };
+  }
+
+	var hmrStr = options.hmr ? [
 		// Hot Module Replacement,
 		"if(module.hot) {",
 		// When the styles change, update the <style> tags
@@ -68,7 +68,7 @@ module.exports.pitch = function (request) {
 		// When the module is disposed, remove the <style> tags
 		"	module.hot.dispose(function() { update(); });",
 		"}"
-	].join("\n");
+	].join("\n") : "";
 
 	return [
 		// Style Loader
@@ -95,6 +95,6 @@ module.exports.pitch = function (request) {
 		"",
 		"if(content.locals) module.exports = content.locals;",
 		"",
-		options.hmr ? hmr : ""
+		hmrStr
 	].join("\n");
 };