--- a/url.js
+++ b/url.js
@@ -6,19 +6,41 @@ var path = require('path');
 
 var loaderUtils = require('loader-utils');
 var validateOptions = require('schema-utils');
+var schema = require('./options.json');
+
+var addStyleUrlPath = path.join(__dirname, "lib", "addStyleUrl.js");
 
 module.exports = function () {};
 
 module.exports.pitch = function (request) {
 	if (this.cacheable) this.cacheable();
 
-	var options = loaderUtils.getOptions(this) || {};
+  var options = loaderUtils.getOptions(this);
 
-	validateOptions(require('./options.json'), options, 'Style Loader (URL)');
+	// The variable is needed, because the function should be inlined.
+	// If is just stored it in options, JSON.stringify will quote
+	// the function and it would be just a string at runtime
+	var insertInto;
 
-	options.hmr = typeof options.hmr === 'undefined' ? true : options.hmr;
+  if (options) {
+    validateOptions(schema, options, 'Style Loader (Useable)');
+    options.hmr = typeof options.hmr === 'undefined' ? true : options.hmr;
 
-	var hmr = [
+    var insertIntoType = typeof options.insertInto;
+    if (insertIntoType === "function") {
+      insertInto = options.insertInto.toString();
+    }
+
+    // We need to check if it a string, or variable will be "undefined"
+    // and the loader crashes
+    else if (insertIntoType === "string") {
+      insertInto = '"' + options.insertInto + '"';
+    }
+  } else {
+    options = { hmr: true };
+  }
+
+	var hmrStr = options.hmr ? [
 		// Hot Module Replacement
 		"if(module.hot) {",
 		"  module.hot.accept(" + loaderUtils.stringifyRequest(this, "!!" + request) + ", function() {",
@@ -27,13 +49,13 @@ module.exports.pitch = function (request) {
 		"",
 		"  module.hot.dispose(function() { update(); });",
 		"}"
-	].join("\n");
+	].join("\n") : "";
 
 	return [
 		// Adds some reference to a CSS file to the DOM by adding a <link> tag
-		"var update = require(" + loaderUtils.stringifyRequest(this, "!" + path.join(__dirname, "lib", "addStyleUrl.js")) + ")(",
+		"var update = require(" + loaderUtils.stringifyRequest(this, "!" + addStyleUrlPath) + ")(",
 		"  require(" + loaderUtils.stringifyRequest(this, "!!" + request) + ")",
 		", " + JSON.stringify(options) + ");",
-		options.hmr ? hmr : ""
+		hmrStr
 	].join("\n");
 };