--- a/useable.js
+++ b/useable.js
@@ -6,34 +6,41 @@ var path = require('path');
 
 var loaderUtils = require("loader-utils");
 var validateOptions = require('schema-utils');
+var schema = require('./options.json');
+
+var addStylesPath = path.join(__dirname, "lib", "addStyles.js");
 
 module.exports = function () {};
 
 module.exports.pitch = function (request) {
 	if (this.cacheable) this.cacheable();
 
-	var options = loaderUtils.getOptions(this) || {};
-
-	validateOptions(require('./options.json'), options, 'Style Loader (Useable)');
-
-	options.hmr = typeof options.hmr === 'undefined' ? true : options.hmr;
+  var options = loaderUtils.getOptions(this);
 
 	// The variable is needed, because the function should be inlined.
 	// If is just stored it in options, JSON.stringify will quote
 	// the function and it would be just a string at runtime
 	var insertInto;
 
-	if (typeof options.insertInto === "function") {
-		insertInto = options.insertInto.toString();
-	}
+  if (options) {
+    validateOptions(schema, options, 'Style Loader (Useable)');
+    options.hmr = typeof options.hmr === 'undefined' ? true : options.hmr;
 
-	// We need to check if it a string, or variable will be "undefined"
-	// and the loader crashes
-	if (typeof options.insertInto === "string") {
-		insertInto = '"' + options.insertInto + '"';
-	}
+    var insertIntoType = typeof options.insertInto;
+    if (insertIntoType === "function") {
+      insertInto = options.insertInto.toString();
+    }
 
-	var hmr = [
+    // We need to check if it a string, or variable will be "undefined"
+    // and the loader crashes
+    else if (insertIntoType === "string") {
+      insertInto = '"' + options.insertInto + '"';
+    }
+  } else {
+    options = { hmr: true };
+  }
+
+	var hmrStr = options.hmr ? [
 		// Hot Module Replacement
 		"if(module.hot) {",
 		"	var lastRefs = module.hot.data && module.hot.data.refs || 0;",
@@ -57,7 +64,7 @@ module.exports.pitch = function (request) {
 		"		}",
 		"	});",
 		"}"
-	].join("\n");
+	].join("\n") : "";
 
 	return [
 		"var refs = 0;",
@@ -72,7 +79,7 @@ module.exports.pitch = function (request) {
 		"",
 		"exports.use = exports.ref = function() {",
 		"	if(!(refs++)) {",
-		"		dispose = require(" + loaderUtils.stringifyRequest(this, "!" + path.join(__dirname, "lib", "addStyles.js")) + ")(content, options);",
+		"		dispose = require(" + loaderUtils.stringifyRequest(this, "!" + addStylesPath) + ")(content, options);",
 		"	}",
 		"",
 		"	return exports;",
@@ -84,6 +91,6 @@ module.exports.pitch = function (request) {
 		"		 dispose = null;",
 		"  }",
 		"};",
-		options.hmr ? hmr : ""
+		hmrStr
 	].join("\n");
 };