This change aims to improve perf by reducing extraneous called to `validateOptions()` & hoisting `require()`.

**What kind of change does this PR introduce?**
perf

**Did you add tests for your changes?**
Yes

**Summary**

This change aims to improve perf by reducing extraneous called to `validateOptions()` & hoisting `require()`.

**Does this PR introduce a breaking change?**
No