@hulkish 2019-01-26
```js
const Benchmark = require('benchmark');

const styleLoader = require('.');

Benchmark.options.minSamples = 100;

const loaderContextWithoutOptions = {
  cacheable() {},
  getOptions() {}
};

console.log('without options');
new Benchmark.Suite()
  .add({
    name: 'styleLoader.oldPitch()',
    fn() {
      styleLoader.oldPitch.call(loaderContextWithoutOptions, 'foo');
    },
  })
  .add({
    name: 'styleLoader.pitch()',
    fn() {
      styleLoader.pitch.call(loaderContextWithoutOptions, 'foo');
    },
  })
  .on('cycle', (event) => {
    console.log(String(event.target));
  })
  .on('complete', function() {
    console.log('Fastest is', this.filter('fastest').map('name'));
  })
  .run();

const options = {
  insertInto() { console.log('foo'); }
};
const loaderContextWithOptions = {
  cacheable() {},
  getOptions() { return options }
}

console.log('with options');
new Benchmark.Suite()
  .add({
    name: 'styleLoader.oldPitch()',
    fn() {
      styleLoader.oldPitch.call(loaderContextWithOptions, 'foo');
    },
  })
  .add({
    name: 'styleLoader.pitch()',
    fn() {
      styleLoader.pitch.call(loaderContextWithOptions, 'foo');
    },
  })
  .on('cycle', (event) => {
    console.log(String(event.target));
  })
  .on('complete', function() {
    console.log('Fastest is', this.filter('fastest').map('name'));
  })
  .run();
```

```sh
~/dev/repos/style-loader$ node benchmark.js
without options
styleLoader.oldPitch() x 37,352 ops/sec ±1.18% (184 runs sampled)
styleLoader.pitch() x 107,537 ops/sec ±1.36% (186 runs sampled)
Fastest is [ 'styleLoader.pitch()' ]
with options
styleLoader.oldPitch() x 39,174 ops/sec ±1.23% (185 runs sampled)
styleLoader.pitch() x 103,644 ops/sec ±1.27% (184 runs sampled)
Fastest is [ 'styleLoader.pitch()' ]
```