@evilebottnawi 2019-07-29
Better handle this case https://github.com/webpack/schema-utils