@evilebottnawi 2019-07-29
We can discuss about this here https://github.com/webpack/schema-utils/issues/45